Move Table out of ManageUsers component
