+++ frontend/src/Dashboard/Table/BulkActionsDropdown.scss
