+++ frontend/src/Dashboard/Table/Table.scss
