+++ frontend/src/Dashboard/Table/TableSearch.scss
